# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Clear Clinic - Cosmetic & Dental Clinic API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [cosmetic, dental, both]
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a service by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # PACKAGES
  /packages:
    get:
      operationId: listPackages
      tags: [packages]
      summary: List packages and offers
      parameters:
        - name: featured
          in: query
          schema:
            type: boolean
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
    post:
      operationId: createPackage
      tags: [packages]
      summary: Create a package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageInput"
      responses:
        "201":
          description: Package created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
  /packages/{id}:
    get:
      operationId: getPackage
      tags: [packages]
      summary: Get a package by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
    patch:
      operationId: updatePackage
      tags: [packages]
      summary: Update a package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
    delete:
      operationId: deletePackage
      tags: [packages]
      summary: Delete a package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # APPOINTMENTS
...
  /testimonials:
    get:
      operationId: listTestimonials
      tags: [testimonials]
      summary: List testimonials
      parameters:
        - name: approved
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of testimonials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"
    post:
      operationId: createTestimonial
      tags: [testimonials]
      summary: Submit a testimonial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialInput"
      responses:
        "201":
          description: Testimonial submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
  /testimonials/{id}:
    patch:
      operationId: updateTestimonial
      tags: [testimonials]
      summary: Approve or update a testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
    delete:
      operationId: deleteTestimonial
      tags: [testimonials]
      summary: Delete a testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # DASHBOARD
...
components:
  schemas:
...
    Service:
      type: object
      required: [id, nameAr, nameEn, category, active]
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        descriptionAr:
          type: ["string", "null"]
        descriptionEn:
          type: ["string", "null"]
        category:
          type: string
          enum: [cosmetic, dental, both]
        imageUrl:
          type: ["string", "null"]
        durationMinutes:
          type: ["integer", "null"]
        priceFrom:
          type: ["number", "null"]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ServiceInput:
      type: object
      required: [nameAr, nameEn, category]
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
        descriptionAr:
          type: string
        descriptionEn:
          type: string
        category:
          type: string
          enum: [cosmetic, dental, both]
        imageUrl:
          type: string
        durationMinutes:
          type: integer
        priceFrom:
          type: number
        active:
          type: boolean
    ServiceUpdate:
      type: object
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
        descriptionAr:
          type: string
        descriptionEn:
          type: string
        category:
          type: string
          enum: [cosmetic, dental, both]
        imageUrl:
          type: string
        durationMinutes:
          type: integer
        priceFrom:
          type: number
        active:
          type: boolean
    Package:
      type: object
      required: [id, nameAr, nameEn, active]
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        descriptionAr:
          type: ["string", "null"]
        descriptionEn:
          type: ["string", "null"]
        originalPrice:
          type: ["number", "null"]
        discountedPrice:
          type: ["number", "null"]
        discountPercent:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
          format: date
        featured:
          type: boolean
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    PackageInput:
      type: object
      required: [nameAr, nameEn]
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
        descriptionAr:
          type: string
        descriptionEn:
          type: string
        originalPrice:
          type: number
        discountedPrice:
          type: number
        discountPercent:
          type: number
        imageUrl:
          type: string
        validUntil:
          type: string
          format: date
        featured:
          type: boolean
        active:
          type: boolean
    PackageUpdate:
      type: object
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
        descriptionAr:
          type: string
        descriptionEn:
          type: string
        originalPrice:
          type: number
        discountedPrice:
          type: number
        discountPercent:
          type: number
        imageUrl:
          type: string
        validUntil:
          type: string
          format: date
        featured:
          type: boolean
        active:
          type: boolean
...
    Testimonial:
      type: object
      required: [id, patientName, rating, approved]
      properties:
        id:
          type: integer
        patientName:
          type: string
        rating:
          type: integer
        commentAr:
          type: ["string", "null"]
        commentEn:
          type: ["string", "null"]
        serviceId:
          type: ["integer", "null"]
        approved:
          type: boolean
        createdAt:
          type: string
          format: date-time
    TestimonialInput:
      type: object
      required: [patientName, rating]
      properties:
        patientName:
          type: string
        rating:
          type: integer
        commentAr:
          type: string
        commentEn:
          type: string
        serviceId:
          type: integer
    TestimonialUpdate:
      type: object
      properties:
        approved:
          type: boolean
        commentAr:
          type: string
        commentEn:
          type: string
...