# lib/api-zod/src/generated/types/packageUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clear Clinic - Cosmetic & Dental Clinic API
 * OpenAPI spec version: 0.1.0
 */

export interface PackageUpdate {
  nameAr?: string;
  nameEn?: string;
  descriptionAr?: string;
  descriptionEn?: string;
  originalPrice?: number;
  discountedPrice?: number;
  discountPercent?: number;
  imageUrl?: string;
  validUntil?: Date;
  featured?: boolean;
  active?: boolean;
}