# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clear Clinic - Cosmetic & Dental Clinic API
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceCategory } from "./serviceCategory";

export interface Service {
  id: number;
  nameAr: string;
  nameEn: string;
  /** @nullable */
  descriptionAr?: string | null;
  /** @nullable */
  descriptionEn?: string | null;
  category: ServiceCategory;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  priceFrom?: number | null;
  active: boolean;
  createdAt?: Date;
}