# lib/api-zod/src/generated/types/serviceUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clear Clinic - Cosmetic & Dental Clinic API
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceUpdateCategory } from "./serviceUpdateCategory";

export interface ServiceUpdate {
  nameAr?: string;
  nameEn?: string;
  descriptionAr?: string;
  descriptionEn?: string;
  category?: ServiceUpdateCategory;
  imageUrl?: string;
  durationMinutes?: number;
  priceFrom?: number;
  active?: boolean;
}